move some python files
