+++ src/asset/lod.py
@@ -1,5 +1,5 @@
-from io_util import *
-from logger import logger
+from util.io_util import *
+from util.logger import logger
 
 class LODSection:
     # material_id: material id
@@ -553,103 +553,3 @@ class LOD:
 
         for v in self.vertices:
             v.lower_buffer()
-
-    '''
-    def embed_data_into_VB(self, bin):
-        size=len(bin)
-        if size>=256:
-            logger.error('The length of author name should be smaller than 256.')
-        bin=b''.join([bin, b'\x00'*(-len(bin)%4)])
-        vb_sample=self.vertices[0].vb
-        vb_size=len(vb_sample)
-        vb2_size=self.influence_size
-        normal=vb_sample[:8]
-        true_vertex_num=len(self.vertices)
-        fake_vb=b''.join([normal, b'\x00'*(vb_size-8)])
-        fake_vb2_id=b'\x00'*(vb2_size//2)
-        for i in range(len(bin)//4):
-            fake_vb2=b''.join([fake_vb2_id, bin[i*4:(i+1)*4], b'\x00'*(vb2_size//2-4)])
-            vertex=Vertex(fake_vb)
-            vertex.vb2=fake_vb2
-            self.vertices.append(vertex)
-        fake_vertex_num=len(self.vertices)-true_vertex_num+1
-
-        bin=[0,13,fake_vertex_num, size]
-        bin=[b.to_bytes(1, 'big') for b in bin]
-        bin=b''.join(bin)
-        fake_vb2=b''.join([fake_vb2_id, bin, b'\x00'*(vb2_size//2-4)])
-        vertex=Vertex(fake_vb)
-        vertex.vb2=fake_vb2
-        self.vertices.append(vertex)
-
-        self.sections[-1].vertex_num+=fake_vertex_num
-        if self.KDI_buffer_size>0:
-            self.KDI_VB=self.KDI_VB+[-1]*fake_vertex_num
-
-        return fake_vertex_num
-
-    def get_metadata(self):
-        v=self.vertices[-1]
-        vb2_size=self.influence_size
-        fake_vb2=list(v.vb2[vb2_size//2:vb2_size//2+4])
-        if fake_vb2[0]!=0 or fake_vb2[1]!=13:
-            return b''
-        fake_vertex_num=fake_vb2[2]
-        size=fake_vb2[3]
-        fake_vertices=self.vertices[-fake_vertex_num:-1]
-        vb_list=[v.vb2[vb2_size//2:vb2_size//2+4] for v in fake_vertices]
-        data = b''.join(vb_list)
-        return data[:size]
-    '''
-
-class PhysicalMesh: #collider or something? low poly mesh.
-    #vertices
-    #bone_id: vertex group? each vertex has a bone id.
-    #faces
-
-    def __init__(self, f):
-        self.offset=f.tell()
-        vertex_num=read_uint32(f)
-        self.vb=f.read(vertex_num*12)
-        #self.vertices=read_vec3_f32_array(f)
-        #vertex_num=len(self.vertices)
-        
-        num = read_uint32(f)
-        check(num, vertex_num, f, 'Parse failed! (StaticMesh:vertex_num)')
-        
-        self.weight_buffer=f.read(num*12)
-        '''
-        self.bone_id=[]
-        self.weight=[]
-        for i in range(num):
-            bone_id=read_uint16_array(f, len=4)
-            self.bone_id.append(bone_id)
-            weight=read_uint8_array(f, len=4)
-            self.weight.append(weight)
-        '''
-
-        face_num=read_uint32(f)
-        self.faces=Face.read_array(f, len=face_num)
-
-    def read(f):
-        return PhysicalMesh(f)
-
-    def write(f, mesh):
-        write_uint32(f, len(mesh.vb)//12)
-        f.write(mesh.vb)
-        #write_vec3_f32_array(f, mesh.vertices, with_length=True)
-        write_uint32(f, len(mesh.weight_buffer)//12)
-        f.write(mesh.weight_buffer)
-        '''
-        for bone_id, weight in zip(mesh.bone_id, mesh.weight):
-            write_uint16_array(f, bone_id)
-            write_uint8_array(f, weight)
-        '''
-        write_uint32(f, len(mesh.faces)//6)
-        Face.write_array(f, mesh.faces)
-
-    def print(self, padding=0):
-        pad=' '*padding
-        logger.log(pad+'Mesh (offset: {})'.format(self.offset))
-        logger.log(pad+'  vertex_num: {}'.format(len(self.vb)//12))
-        logger.log(pad+'  face_num: {}'.format(len(self.faces)))
