+++ src/asset/uexp.py
@@ -2,13 +2,13 @@
 import os
 
 #my libs
-from io_util import *
-from logger import logger
-from cipher import Cipher
-
-from skeletal_mesh import SkeletalMesh
+from util.io_util import *
+from util.logger import logger
+from util.cipher import Cipher
+from asset.skeletal_mesh import SkeletalMesh
+from asset.static_mesh import StaticMesh
 #from unknown_block import unknown
-from uasset import Uasset
+from asset.uasset import Uasset
 
 class MeshUexp:
 
@@ -31,8 +31,8 @@ class MeshUexp:
         self.imports = self.uasset.imports
         self.ff7r = self.uasset.ff7r
         self.skeletal = self.uasset.skeletal
-        if not self.skeletal:
-            logger.error('Not skeletal mesh. ({})'.format(file))
+        #if not self.skeletal:
+        #    logger.error('Not skeletal mesh. ({})'.format(file))
         logger.log('FF7R: {}'.format(self.ff7r))
 
         logger.log('')
@@ -50,15 +50,23 @@ class MeshUexp:
                     export.read_uexp(f)
                     
                 else:
-                    if export.id==-1 and self.skeletal:
-                        self.skeletalmesh=SkeletalMesh.read(f, self.ff7r, self.name_list, self.imports)                        
-                        self.unknown2=f.read(export.offset+export.size-f.tell()-self.uasset.size)
+                    if export.id==-1:
+                        if self.skeletal:
+                            self.skeletalmesh=SkeletalMesh.read(f, self.ff7r, self.name_list, self.imports)                        
+                            self.unknown2=f.read(export.offset+export.size-f.tell()-self.uasset.size)
+                        else:
+                            logger.error('Not skeltal mesh.')
+                            self.staticmesh=StaticMesh.read(f, self.ff7r, self.name_list, self.imports)
 
             #footer
             offset = f.tell()
             size = get_size(f)
-            self.meta=f.read(size-offset-4)
-            self.author = Cipher.decrypt(self.meta)
+            if self.skeletal:
+                self.meta=f.read(size-offset-4)
+                self.author = Cipher.decrypt(self.meta)                
+            else:
+                self.author=''
+
             if self.author!='':
                 print('Author: {}'.format(self.author))
             self.foot=f.read()
@@ -73,10 +81,14 @@ class MeshUexp:
                     export.write_uexp(f)
                     size=export.size
                 else:
-                    if export.id==-1 and self.skeletal:
-                        SkeletalMesh.write(f, self.skeletalmesh)
-                        f.write(self.unknown2)
+                    if export.id==-1:
+                        if self.skeletal:
+                            SkeletalMesh.write(f, self.skeletalmesh)
+                            f.write(self.unknown2)
+                        else:
+                            StaticMesh.write(f, self.staticmesh)
                         size=f.tell()-offset
+
                 export.update(size, offset+self.uasset.size)
 
             f.write(self.meta)
@@ -87,6 +99,8 @@ class MeshUexp:
     def remove_LODs(self):
         if self.skeletal:
             self.skeletalmesh.remove_LODs()
+        else:
+            logger.error('Unsupported method for static mesh')
 
     def import_LODs(self, mesh_uexp, only_mesh=False, dont_remove_KDI=False):
         if not mesh_uexp.skeletal:
@@ -94,17 +108,22 @@ class MeshUexp:
 
         if self.skeletal:
             self.skeletalmesh.import_LODs(mesh_uexp.skeletalmesh, only_mesh=only_mesh, dont_remove_KDI=dont_remove_KDI)
+        else:
+            logger.error('Unsupported method for static mesh')
 
     def remove_KDI(self):
         if self.skeletal:
             self.skeletalmesh.remove_KDI()
+        else:
+            logger.error('Unsupported method for static mesh')
 
     def dump_buffers(self, save_folder):
         if self.skeletal:
             self.skeletalmesh.dump_buffers(save_folder)
+        else:
+            logger.error('Unsupported method for static mesh')
 
     def embed_string(self, string):
-        #self.skeletalmesh.embed_data_into_VB(bin)
         self.author=string
         self.meta=Cipher.encrypt(string)
         logger.log('A string has been embedded into uexp.', ignore_verbose=True)
@@ -113,4 +132,3 @@ class MeshUexp:
 
     def get_author(self):
         return self.author
-        #return self.skeletalmesh.get_metadata()
