+++ src/asset/unknown_block.py
@@ -1,5 +1,5 @@
-from io_util import *
-from logger import logger
+from util.io_util import *
+from util.logger import logger
 
 class unknown_sub:
     UNK=b'\x01\x00\x00\x70\x42\x03\x03\x03\x03\x00\x00\xC8\x42\x04\x00\x00\x00'
